Align ArchitectureImplementation naming with the Base API specification
Rename and retype the remaining methods to match the specification. The idle AICLK
frequency moves onto the interface, with the concrete devices delegating to it, and
the busy frequency stays next to it.

+++ device/tt_device/tt_sim_tt_device.cpp
@@ -294,7 +294,7 @@ bool TTSimTTDevice::special_dram_read(void* mem_ptr, tt_xy_pair core, uint64_t a
 void TTSimTTDevice::assert_risc_reset(CoreCoord core, const RiscType selected_riscs) {
     std::lock_guard<std::recursive_mutex> lock(device_lock);
     log_debug(tt::LogEmulationDriver, "Sending 'assert_risc_reset' signal for risc_type {}", selected_riscs);
-    uint32_t soft_reset_addr = architecture_impl_->get_tensix_soft_reset_addr();
+    uint64_t soft_reset_addr = architecture_impl_->get_tensix_soft_reset_addr();
     uint32_t soft_reset_update = architecture_impl_->get_soft_reset_reg_value(selected_riscs);
     if (libttsim_pci_device_id == 0xFEED) {  // QSR
         uint64_t reset_value;
@@ -313,7 +313,7 @@ void TTSimTTDevice::assert_risc_reset(CoreCoord core, const RiscType selected_ri
 void TTSimTTDevice::deassert_risc_reset(CoreCoord core, const RiscType selected_riscs, bool staggered_start) {
     std::lock_guard<std::recursive_mutex> lock(device_lock);
     log_debug(tt::LogEmulationDriver, "Sending 'deassert_risc_reset' signal for risc_type {}", selected_riscs);
-    uint32_t soft_reset_addr = architecture_impl_->get_tensix_soft_reset_addr();
+    uint64_t soft_reset_addr = architecture_impl_->get_tensix_soft_reset_addr();
     uint32_t soft_reset_update = architecture_impl_->get_soft_reset_reg_value(selected_riscs);
 
     if (libttsim_pci_device_id == 0xFEED) {  // QSR
